Enable Trace API
Update the CIRCT phase to support Chisel's Trace API.  This is an API
that allows a user to add a TraceNameAnnotation that will be updated and
output by a FIRRTL compiler.  The intent of this API is that it gives a
user the ability to find the final FIRRTL Target for the original
annotated component.  This is intentionally a low-level API that enables
building more complex APIs, e.g., TCL emission, without requiring custom
pass injection into the FIRRTL compiler.

This requires a few modifications to internal to find/load an output
annotation file from CIRCT.  This uses a new "-output-annotation-file"
option to CIRCT specific to this API.  When in "-split-verilog" output
mode, the output annotation is found and read.  When in single file
output mode, the output single-file string is parsed to extract the
annotation section.

This patch requires using a version of CIRCT that includes changes from
https://github.com/llvm/circt/pull/4065.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>