Remove CIRCT "Handover" Functionality
Remove the "CIRCTHandover" command line option/annotation and all code
related to it.  This is an unused feature that enables running the
Scala-based FIRRTL Compiler (SFC) up to a specific "handover
point" (CHIRRTL, High FIRRTL, etc.) and then switching to the MLIR-based
FIRRTL Compiler (MFC), i.e., CIRCT.  This was originally envisioned as a
testing functionality while major parts of the MFC were brought online.
E.g., this was very useful before MFC had a LowerTypes pass.

More importantly, this isn't a workflow that we want to support and we
want chisel-circt to be fully on the MFC and not straddling both the SFC
and MFC.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>